Documentation and cleanup
Add a new README.md
Add a new example manifest example/manifest/sample_secrets_manifest.yaml
