+++ .github/workflows/ci.yml
@@ -12,10 +12,10 @@ jobs:
     name: build and test
     runs-on: ubuntu-18.04
     steps:
-      - name: Set up Go 1.15.7
+      - name: Golang setup
         uses: actions/setup-go@v1
         with:
-          go-version: 1.15.7
+          go-version: 1.15
         id: go
 
       - name: Check out code into the Go module directory
@@ -28,8 +28,8 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-go-
 
-      - name: vet
+      - name: golang vet
         run: go vet ./...
 
-      - name: unit test
+      - name: golang  unit test
         run: go test ./... -cover -race
