+++ .github/workflows/manual_release.yml
@@ -1,19 +1,9 @@
-# This is a basic workflow that is manually triggered
-
-name: Manual release workflow
-
-# Controls when the action will run. Workflow runs when manually triggered using the UI
-# or API.
+name: Manual release
 on:
   workflow_dispatch:
-    # Inputs the workflow accepts.
     inputs:
       version:
-        # Friendly description to be shown in the UI instead of 'name'
         description: 'Version for the released image'
-        # Default value if no value is explicitly provided
-        #default: 'World'
-        # Input has to be provided for the workflow to run
         required: true
 
 jobs:
@@ -23,7 +13,7 @@ jobs:
     env:
       SERVICE: secretsfetcher
     steps:
-      - name: Set up Go 1.15
+      - name: Golang setup
         uses: actions/setup-go@v1
         with:
           go-version: 1.15
