+++ .github/workflows/release.yml
@@ -12,7 +12,7 @@ jobs:
     name: build and test
     runs-on: ubuntu-18.04
     steps:
-      - name: Set up Go 1.15
+      - name: Golang setup
         uses: actions/setup-go@v1
         with:
           go-version: 1.15
