+++ cmd/aws.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"io/ioutil"
-	"log"
 	"os"
 	"path"
 	"strings"
@@ -11,7 +10,6 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 )
 
 const (
@@ -22,14 +20,7 @@ const (
 var (
 	manifestFile string
 	outputFolder string
-	//tagFilters   map[string]string
-	//secretPrefix string
 )
-
-// func initConfig() {
-// 	//pf.StringVarP(&manifestFile, "manifest", "m", "", "secrets manifest file")
-// 	//cobra.MarkFlagRequired(pf, "manifest")
-// }
 
 // versionCmd represents the version command
 var awsCmd = &cobra.Command{
@@ -153,40 +144,4 @@ func init() {
 	//awsCmd.MarkFlagRequired("manifest")
 
 	rootCmd.AddCommand(awsCmd)
-}
-
-func initLog(logLevel string) *zap.Logger {
-	level := zapcore.InfoLevel
-	if err := level.Set(logLevel); err != nil {
-		log.Fatalf("could not set zap log level to: \"%s\" \n", logLevel)
-	}
-
-	config := &zap.Config{
-		Encoding:         "json",
-		Level:            zap.NewAtomicLevelAt(level),
-		OutputPaths:      []string{"stdout"},
-		ErrorOutputPaths: []string{"stdout"},
-		EncoderConfig: zapcore.EncoderConfig{
-			MessageKey: "message",
-
-			LevelKey:    "severity",
-			EncodeLevel: zapcore.CapitalLevelEncoder,
-
-			TimeKey:    "time",
-			EncodeTime: zapcore.RFC3339NanoTimeEncoder,
-
-			CallerKey:      "src",
-			EncodeCaller:   zapcore.ShortCallerEncoder,
-			EncodeDuration: zapcore.SecondsDurationEncoder,
-
-			StacktraceKey: "stk",
-		},
-	}
-
-	zl, err := config.Build()
-	if err != nil {
-		log.Fatalf("failed to build zap logger. Error: %s", err)
-	}
-	return zl
-
 }
