+++ example/manifest/sample_secrets_manifest.yaml
@@ -1,0 +1,15 @@
+# Ref: https://github.com/aws/secrets-store-csi-driver-provider-aws#secretproviderclass-options
+provider: aws
+secretObjects:
+  - objectName: "arn:aws:secretsmanager:ap-southeast-2:111122223333:secret:my-secret-prefix/secret-path/secret-name-1a2b3c"
+    objectType: "secretsmanager"
+    objectVersion: "9b324964-c0ff-4adf-aeda-ad403efdaeb9" # [OPTIONAL] object versions, default to latest if empty
+
+
+  - objectName: "my-secret-prefix/secret-path/secret-name-2"
+    objectType: "secretsmanager"
+    #objectVersionLabel: "AWSCURRENT" # [OPTIONAL] object version stage, default to latest if empty
+
+region: ap-southeast-2
+#pathTranslation: "False"
+pathTranslation: "$"
